Add breadboard leg ends
Allows the creator / printer to add legs to the end of an particularly
thin board to keep it off of the desk / table / floor / whatever.

Originally used to pick a fairly cheap breadboard up off of the desk
because it had screws that ran through a thin metal plate, but adhesive
feet would not stick to it very well.

Signed-off-by: Michael Bergeron <mike@michaelbergeron.com>